refactor: inline lend routes
