--- apps/main/src/app/lend/[network]/integrations/page.tsx
@@ -1,8 +1,0 @@
-import Integrations from '@/lend/components/PageIntegrations/Page'
-import type { NetworkUrlParams } from '@/lend/types/lend.types'
-import { useParams } from '@ui-kit/hooks/router'
-
-export default function Component() {
-  const params = useParams<NetworkUrlParams>()
-  return <Integrations {...params} />
-}
