--- apps/main/src/app/lend/[network]/markets/[market]/create/[[...formType]]/page.tsx
@@ -1,8 +1,0 @@
-import LoanCreate from '@/lend/components/PageLoanCreate/Page'
-import type { MarketUrlParams } from '@/lend/types/lend.types'
-import { useParams } from '@ui-kit/hooks/router'
-
-export default function Component() {
-  const params = useParams<MarketUrlParams>()
-  return <LoanCreate {...params} />
-}
