--- apps/main/src/app/lend/[network]/markets/[market]/manage/[[...formType]]/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import LoanManage from '@/lend/components/PageLoanManage/Page'
-import type { MarketUrlParams } from '@/lend/types/lend.types'
-import { useParams } from '@ui-kit/hooks/router'
-
-export default function Component() {
-  const params = useParams<MarketUrlParams>()
-  return <LoanManage {...params} />
-}
