--- apps/main/src/app/llamalend/[network]/markets/page.tsx
@@ -1,5 +1,0 @@
-import { LlamaMarketsPage } from '@/llamalend/PageLlamaMarkets/Page'
-
-export default function Component() {
-  return <LlamaMarketsPage />
-}
