+++ apps/main/src/lend/components/PageIntegrations/Page.tsx
@@ -8,10 +8,11 @@ import type { NetworkUrlParams } from '@/lend/types/lend.types'
 import { ExternalLink } from '@ui/Link'
 import Spinner, { SpinnerWrapper } from '@ui/Spinner'
 import { breakpoints } from '@ui/utils/responsive'
-import { useSearchParams } from '@ui-kit/hooks/router'
+import { useSearchParams, useParams } from '@ui-kit/hooks/router'
 import { Trans } from '@ui-kit/lib/i18n'
 
-const Page = (params: NetworkUrlParams) => {
+const Page = () => {
+  const params = useParams<NetworkUrlParams>()
   const searchParams = useSearchParams()
   const rChainId = networksIdMapper[params.network]
