+++ apps/main/src/routes/lend.routes.tsx
@@ -1,17 +1,42 @@
-import { lazy } from 'react'
-import LendLayout from '@/app/lend/layout'
+'use client'
+import '@/global-extensions'
+import PageIntegrations from '@/lend/components/PageIntegrations/Page'
+import PageLoanCreate from '@/lend/components/PageLoanCreate/Page'
+import PageLoanManage from '@/lend/components/PageLoanManage/Page'
+import PageMarkets from '@/lend/components/PageMarketList/Page'
+import PageVault from '@/lend/components/PageVault/Page'
+import networks, { networksIdMapper } from '@/lend/networks'
+import useStore from '@/lend/store/useStore'
+import type { UrlParams } from '@/lend/types/lend.types'
+import Skeleton from '@mui/material/Skeleton'
 import { createRoute, Outlet } from '@tanstack/react-router'
+import { useParams } from '@ui-kit/hooks/router'
+import { useHydration } from '@ui-kit/hooks/useHydration'
+import { useRedirectToEth } from '@ui-kit/hooks/useRedirectToEth'
+import { useGasInfoAndUpdateLib } from '@ui-kit/lib/model/entities/gas-info'
+import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
+import { Disclaimer } from '@ui-kit/widgets/Disclaimer'
 import { rootRoute } from './root.routes'
 import { redirectTo } from './util'
+
+const { MinHeight } = SizesAndSpaces
+
+function LendLayout() {
+  const { network: networkId = 'ethereum' } = useParams<Partial<UrlParams>>()
+  const chainId = networksIdMapper[networkId]
+  const hydrate = useStore((s) => s.hydrate)
+  const isHydrated = useHydration('llamaApi', hydrate, chainId)
+
+  useRedirectToEth(networks[chainId], networkId, isHydrated)
+  useGasInfoAndUpdateLib({ chainId, networks })
+
+  return isHydrated && <Outlet />
+}
 
 export const lendLayoutRoute = createRoute({
   getParentRoute: () => rootRoute,
   path: 'lend',
-  component: () => (
-    <LendLayout>
-      <Outlet />
-    </LendLayout>
-  ),
+  component: LendLayout,
 })
 
 const layoutProps = { getParentRoute: () => lendLayoutRoute }
@@ -19,7 +44,7 @@ const layoutProps = { getParentRoute: () => lendLayoutRoute }
 export const lendRoutes = lendLayoutRoute.addChildren([
   createRoute({
     path: '/',
-    component: lazy(() => import('../app/lend/page')),
+    component: () => <Skeleton width="100%" height={MinHeight.pageContent} />,
     head: () => ({
       meta: [{ title: 'Lend - Curve' }],
     }),
@@ -32,7 +57,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/disclaimer',
-    component: lazy(() => import('../app/lend/[network]/disclaimer/page')),
+    component: () => <Disclaimer currentApp="lend" />,
     head: () => ({
       meta: [{ title: 'Risk Disclaimer - Curve Lend' }],
     }),
@@ -40,7 +65,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/integrations',
-    component: lazy(() => import('../app/lend/[network]/integrations/page')),
+    component: PageIntegrations,
     head: () => ({
       meta: [{ title: 'Integrations - Curve Lend' }],
     }),
@@ -48,7 +73,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/markets',
-    component: lazy(() => import('../app/lend/[network]/markets/page')),
+    component: PageMarkets,
     head: () => ({
       meta: [{ title: 'Markets - Curve Lend' }],
     }),
@@ -56,7 +81,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/markets/$market/create/$formType',
-    component: lazy(() => import('../app/lend/[network]/markets/[market]/create/[[...formType]]/page')),
+    component: PageLoanCreate,
     head: ({ params }) => ({
       meta: [{ title: `${params.market} | Create Loan - Curve Lend` }],
     }),
@@ -64,7 +89,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/markets/$market/create',
-    component: lazy(() => import('../app/lend/[network]/markets/[market]/create/[[...formType]]/page')),
+    component: PageLoanCreate,
     head: ({ params }) => ({
       meta: [{ title: `${params.market} | Create Loan - Curve Lend` }],
     }),
@@ -72,7 +97,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/markets/$market/manage/$formType',
-    component: lazy(() => import('../app/lend/[network]/markets/[market]/manage/[[...formType]]/page')),
+    component: PageLoanManage,
     head: ({ params }) => ({
       meta: [{ title: `${params.market} | Manage Loan - Curve Lend` }],
     }),
@@ -80,7 +105,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/markets/$market/manage',
-    component: lazy(() => import('../app/lend/[network]/markets/[market]/manage/[[...formType]]/page')),
+    component: PageLoanManage,
     head: ({ params }) => ({
       meta: [{ title: `${params.market} | Manage Loan - Curve Lend` }],
     }),
@@ -88,7 +113,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/markets/$market/vault/$formType',
-    component: lazy(() => import('../app/lend/[network]/markets/[market]/vault/[[...formType]]/page')),
+    component: PageVault,
     head: ({ params }) => ({
       meta: [{ title: `${params.market} | Supply - Curve Lend` }],
     }),
@@ -96,7 +121,7 @@ export const lendRoutes = lendLayoutRoute.addChildren([
   }),
   createRoute({
     path: '$network/markets/$market/vault',
-    component: lazy(() => import('../app/lend/[network]/markets/[market]/vault/[[...formType]]/page')),
+    component: PageVault,
     head: ({ params }) => ({
       meta: [{ title: `${params.market} | Supply - Curve Lend` }],
     }),
